passage en modulo 2

+++ include/affichage.h
@@ -5,7 +5,7 @@ void clearTerminal();
 
 void Affichage();
 
-void LireClePublique(char *nomfichier, int taille, float matrix[taille][taille]);
+void LireClePublique(char *nomfichier, int taille, int matrix[taille][taille]);
 
 int nbligne(char * nomFichier);
