+++ include/keygen.h
@@ -4,13 +4,13 @@
 
 #include <stdio.h>
 
-void EcrireMatrice(FILE * fichier, int taille1, int taille2, float matrix[taille1][taille2]);
+void EcrireMatrice(FILE * fichier, int taille1, int taille2, int matrix[taille1][taille2]);
 
 void KeyGen(int size, int redondance);
 
-void GenerateP(int size, float matrix[size][size], int nb_permutation); //génére des matrices de permutations
+void GenerateP(int size, int matrix[size][size], int nb_permutation); //génére des matrices de permutations
 
-void GenerateS(int size, float matrix[size][size]);
+void GenerateS(int size, int matrix[size][size]);
 
-void GenerateG(int size, int redondance, float matrix[size][size+redondance], int itermax);
+void GenerateG(int size, int redondance, int matrix[size][size+redondance], int itermax);
 #endif
